requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
psycopg2
spacy
langdetect
rq-scheduler
redis
rq
websockets
pywin32; sys_platform == "win32"